* master: (65 commits)
  Tweak build destinations
  Implement maintenance-mode actions
  Drop vacuum freeze charm defaults
  Correctly depart conversation in syslog relation
  Fix syslog relation departed hook
  Drop support for PostgreSQL 9.3, 9.4 and Trusty
  Inhibit creation of initial database cluster on package install
  Allow pgdg setting to be changed post-install
  Blacken
  Revert unnecessary code move
  Remove PATH manipulation from the Makefile
  upgrade-series support, resurrect unit tests
  Handle PGDG signing key change
  Handle PGDG archive addition better
  Ensure missing pgkillidle.py is replaced
  blacken
  Blacken with 120 char line length
  OpenStack v3 credentials for WAL-E 1.1.1
  Builds for publishing should use --no-local-layers
  Allow 11 & 12 as valid version numbers in config
  ...

+++ hooks/relations/syslog/provides.py
@@ -23,48 +23,50 @@ from charms.reactive import hook, not_unless, when
 
 
 class SyslogProvides(reactive.relations.RelationBase):
-
-    @hook('{provides:syslog}-relation-changed')
+    @hook("{provides:syslog}-relation-changed")
     def changed(self):
-        self.set_state('{relation_name}.available')
+        self.set_state("{relation_name}.available")
 
-    @hook('{provides:syslog}-relation-departed')
+    @hook("{provides:syslog}-relation-departed")
     def departed(self):
-        path = self.get_local('path')
-        if os.path.exists(path):
+        path = self._rsyslog_conf_path(hookenv.remote_unit())
+        if path and os.path.exists(path):
             os.remove(path)
-            reactive.set_state('syslog.needs_restart')
-        self.remove_state('{relation_name}.available')
-        self.depart()
+            reactive.set_state("syslog.needs_restart")
+        self.remove_state("{relation_name}.available")
+        self.conversation().depart()
 
     def _rsyslog_conf_path(self, remote_unit):
         # Use both the local unit and remote unit in the config file
         # path to avoid conflicts with subordinates.
-        rsyslog_conf_dir = '/etc/rsyslog.d'
-        local = hookenv.local_unit().replace('/', '_')
-        remote = remote_unit.replace('/', '_')
-        return os.path.join(rsyslog_conf_dir,
-                            'juju-{}-{}.conf'.format(local, remote))
+        rsyslog_conf_dir = "/etc/rsyslog.d"
+        local = hookenv.local_unit().replace("/", "_")
+        remote = remote_unit.replace("/", "_")
+        return os.path.join(rsyslog_conf_dir, "juju-{}-{}.conf".format(local, remote))
 
-    @not_unless('{provides:syslog}.available')
+    @not_unless("{provides:syslog}.available")
     def configure(self, programname):
         for conv in self.conversations():
             remote_unit = conv.scope
-            remote_addr = conv.get_remote('private-address')
+            remote_addr = conv.get_remote("private-address")
 
-            if reactive.helpers.data_changed('syslog.{}'.format(remote_unit),
-                                             (remote_unit, remote_addr)):
+            if reactive.helpers.data_changed("syslog.{}".format(remote_unit), (remote_unit, remote_addr)):
                 path = self._rsyslog_conf_path(remote_unit)
 
-                templating.render('rsyslog_forward.conf', path,
-                                  dict(local_unit=hookenv.local_unit(),
-                                       remote_unit=remote_unit,
-                                       remote_addr=remote_addr,
-                                       programname=programname))
-                reactive.set_state('syslog.needs_restart')
+                templating.render(
+                    "rsyslog_forward.conf",
+                    path,
+                    dict(
+                        local_unit=hookenv.local_unit(),
+                        remote_unit=remote_unit,
+                        remote_addr=remote_addr,
+                        programname=programname,
+                    ),
+                )
+                reactive.set_state("syslog.needs_restart")
 
 
-@when('syslog.needs_restart')
+@when("syslog.needs_restart")
 def restart_rsyslog():
-    host.service_restart('rsyslog')
-    reactive.remove_state('syslog.needs_restart')
+    host.service_restart("rsyslog")
+    reactive.remove_state("syslog.needs_restart")
